+++ src/lib.rs
@@ -8,4 +8,4 @@ mod primitives;
 
 pub use error::KikaError;
 pub use predicates::{Orientation, Sign, incircle, insphere, orient2d, orient3d};
-pub use primitives::{Point2, Point3};
+pub use primitives::{Point2, Point3, Vector2, Vector3};