+++ src/primitives/mod.rs
@@ -1,5 +1,9 @@
 mod point2;
 mod point3;
+mod vector2;
+mod vector3;
 
 pub use point2::Point2;
 pub use point3::Point3;
+pub use vector2::Vector2;
+pub use vector3::Vector3;